Move reading part of StringTable to analyzeme crate.
