+++ analyzeme/src/lib.rs
@@ -17,11 +17,13 @@ mod event;
 mod lightweight_event;
 mod profiling_data;
 mod stack_collapse;
-mod timestamp;
+mod stringtable;
 pub mod testing_common;
+mod timestamp;
 
 pub use crate::event::Event;
 pub use crate::lightweight_event::LightweightEvent;
 pub use crate::profiling_data::{ProfilingData, ProfilingDataBuilder};
 pub use crate::stack_collapse::collapse_stacks;
+pub use crate::stringtable::{StringRef, StringTable};
 pub use crate::timestamp::Timestamp;
